Rename 'utilities' to 'utils' to be consistent with styles
