+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import { AppProps } from "next/app";
 import { useEffect } from "react";
 import "../styles/globals.sass";
-import { changeTheme } from "../utilities";
+import { changeTheme } from "../utils";
 
 function MyApp({ Component, pageProps }: AppProps) {
   useEffect(() => {
